requirements.txt
paramiko
pyvmomi
streamlit
scapy
requests
python-nmap
pandas
